# Conflicts:
#	app/src/main/java/com/tosslab/jandi/app/network/models/ResMessages.java
#	app/src/main/java/com/tosslab/jandi/app/push/model/JandiInterfaceModel.java
#	app/src/main/java/com/tosslab/jandi/app/services/socket/JandiSocketServiceModel.java
#	app/src/main/java/com/tosslab/jandi/app/ui/account/model/AccountHomeModel.java
#	app/src/main/java/com/tosslab/jandi/app/ui/account/presenter/AccountHomePresenterImpl.java
#	app/src/main/java/com/tosslab/jandi/app/ui/maintab/mypage/MyPageFragment.java
#	app/src/main/java/com/tosslab/jandi/app/ui/maintab/mypage/view/MyPageView.java
#	app/src/main/java/com/tosslab/jandi/app/ui/message/v2/MessageListV2Presenter.java
#	app/src/main/java/com/tosslab/jandi/app/ui/share/multi/presenter/MultiSharePresenterImpl.java
#	app/src/main/java/com/tosslab/jandi/app/ui/share/presenter/image/ImageSharePresenterImpl.java
#	app/src/main/java/com/tosslab/jandi/app/ui/share/presenter/text/TextSharePresenterImpl.java

+++ app/src/main/java/com/tosslab/jandi/app/ui/message/v2/MessageListV2Presenter.java
@@ -1125,7 +1125,7 @@ public class MessageListV2Presenter {
         view.refreshMessages();
     }
 
-    public void updateStarredOfMessage(int messageId, boolean starred) {
+    public void updateStarredOfMessage(long messageId, boolean starred) {
         int index = adapterModel.indexByMessageId(messageId);
         adapterModel.modifyStarredStateByPosition(index, starred);
         view.refreshMessages();
